perf: reduce per-render allocations in GestureBoard
- Memoize the container style array used by GestureHandlerRootView
  and the inner View; previously each render created two new arrays
  and two new { width } objects, defeating downstream React.memo on
  any wrapper that shallow-compares style props
- Memoize the EffectParams object passed to SkiaBoard; without this
  every render produced a new object reference, breaking
  SkiaBoard's React.memo
- Extract findKingSquare to src/helpers/find-king-square.ts. The
  same body was duplicated in move-executor.ts and gesture-board.tsx;
  both now share the helper

Behavior unchanged.
