+++ src/state/move-executor.ts
@@ -7,6 +7,7 @@ import {
   getChessboardState,
   ChessboardState,
 } from '../helpers/get-chessboard-state';
+import { findKingSquare } from '../helpers/find-king-square';
 
 export type MoveResult = {
   move: Move;
@@ -21,21 +22,6 @@ type MoveCallbacks = {
     color: 'w' | 'b';
     complete: (piece: PieceSymbol) => void;
   }) => void;
-};
-
-const findKingSquare = (chess: Chess, color: 'w' | 'b'): Square | null => {
-  const board = chess.board();
-  for (let row = 0; row < 8; row++) {
-    for (let col = 0; col < 8; col++) {
-      const piece = board[row][col];
-      if (piece && piece.type === 'k' && piece.color === color) {
-        const colChar = String.fromCharCode('a'.charCodeAt(0) + col);
-        const rowNum = 8 - row;
-        return `${colChar}${rowNum}` as Square;
-      }
-    }
-  }
-  return null;
 };
 
 export const createMoveExecutor = (
